Fix: Int/Long type comparison in getKnowledgePointById

+++ app/src/main/java/com/henan/learning/data/local/SimpleDataStore.kt
@@ -38,7 +38,7 @@ class SimpleDataStore(context: Context) {
     }
     
     suspend fun getKnowledgePointById(id: Long): KnowledgePoint? {
-        return _knowledgePoints.value.find { it.id == id }
+        return _knowledgePoints.value.find { it.id.toLong() == id }
     }
     
     // Learning Progress
